+++ packages/date/src/datepicker/datepickerTypes.ts
@@ -1,6 +1,6 @@
 import { LmFormContainerBaseTypes } from '@tamagui-extras/form'
 import { OnDatesChangeProps, UseMonthProps } from '@datepicker-react/hooks'
-import { ButtonProps, InputProps } from 'tamagui'
+import { ButtonProps, InputProps, ThemeableStackProps } from 'tamagui'
 import { LmPopoverProps } from '@tamagui-extras/core'
 
 export type LmDatepickerProps = LmFormContainerBaseTypes & {
@@ -14,5 +14,6 @@ export type LmDatepickerProps = LmFormContainerBaseTypes & {
   fullWidth?: boolean
   inputProps?: InputProps
   buttonProps?: ButtonProps
+  containerProps?: ThemeableStackProps
   popoverProps?: Omit<LmPopoverProps, 'trigger'>
 }