+++ packages/date/types/datepicker/datepickerTypes.d.ts
@@ -1,6 +1,6 @@
 import { LmFormContainerBaseTypes } from '@tamagui-extras/form';
 import { OnDatesChangeProps, UseMonthProps } from '@datepicker-react/hooks';
-import { ButtonProps, InputProps } from 'tamagui';
+import { ButtonProps, InputProps, ThemeableStackProps } from 'tamagui';
 import { LmPopoverProps } from '@tamagui-extras/core';
 export type LmDatepickerProps = LmFormContainerBaseTypes & {
     startDate?: Date | null;
@@ -13,6 +13,7 @@ export type LmDatepickerProps = LmFormContainerBaseTypes & {
     fullWidth?: boolean;
     inputProps?: InputProps;
     buttonProps?: ButtonProps;
+    containerProps?: ThemeableStackProps;
     popoverProps?: Omit<LmPopoverProps, 'trigger'>;
 };
 //# sourceMappingURL=datepickerTypes.d.ts.map