+++ packages/form/types/LmRadioGroup.d.ts
@@ -9,5 +9,5 @@ export type LmRadioGroupProps = RadioGroupProps & LmFormContainerBaseTypes & {
     name?: string;
     spaceItem?: SpaceTokens;
 };
-export declare function LmRadioGroup({ options, required, error, helperText, helperTextProps, label, labelInline, labelProps, size, spaceItem, ...rest }: LmRadioGroupProps): import("react/jsx-runtime").JSX.Element;
+export declare function LmRadioGroup({ options, required, error, helperText, helperTextProps, label, labelInline, labelProps, size, spaceItem, containerProps, ...rest }: LmRadioGroupProps): import("react/jsx-runtime").JSX.Element;
 //# sourceMappingURL=LmRadioGroup.d.ts.map