+++ packages/form/types/LmSlider.d.ts
@@ -7,5 +7,5 @@ export type LmSliderProps = SliderProps & LmFormContainerBaseTypes & {
     thumbProps?: SliderThumbProps;
     showValue?: boolean;
 };
-export declare function LmSlider({ thumbProps, color, defaultValue, label, labelProps, error, required, helperText, helperTextProps, labelInline, colorActiveOnly, showValue, ...sliderProps }: LmSliderProps): import("react/jsx-runtime").JSX.Element;
+export declare function LmSlider({ thumbProps, color, defaultValue, label, labelProps, error, required, helperText, helperTextProps, labelInline, colorActiveOnly, showValue, containerProps, ...sliderProps }: LmSliderProps): import("react/jsx-runtime").JSX.Element;
 //# sourceMappingURL=LmSlider.d.ts.map