+++ packages/form/types/formContainerTypes.d.ts
@@ -1,4 +1,4 @@
-import { LabelProps, ParagraphProps } from 'tamagui';
+import { LabelProps, ParagraphProps, ThemeableStackProps } from 'tamagui';
 export type LmFormContainerBaseTypes = {
     label?: string;
     labelProps?: Omit<LabelProps, 'htmlFor'>;
@@ -7,5 +7,6 @@ export type LmFormContainerBaseTypes = {
     helperTextProps?: ParagraphProps;
     required?: boolean;
     error?: boolean;
+    containerProps?: ThemeableStackProps;
 };
 //# sourceMappingURL=formContainerTypes.d.ts.map