+++ packages/app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # app
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/date@1.7.0
+  - @tamagui-extras/form@1.7.0
+  - @tamagui-extras/rich-text@1.5.1
+
 ## 0.0.18
 
 ### Patch Changes