+++ packages/form/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tamagui-extras/form
 
+## 1.7.0
+
+### Minor Changes
+
+- add containerProps to various components
+
 ## 1.6.0
 
 ### Minor Changes